refactor($output): add message handler
