+++ src/main/java/racinggame/GamePlayer.java
@@ -5,6 +5,9 @@ import racinggame.domain.Cars;
 import racinggame.domain.TryCount;
 
 public class GamePlayer {
+    Cars cars;
+    TryCount tryCount;
+
     public void play(Cars cars, TryCount tryCount) {
         for(int i = 0; i < tryCount.getTryCount(); i++) {
             cars.forward();
@@ -12,4 +15,5 @@ public class GamePlayer {
             Console.printLine("result: " + cars.toString());
         }
     }
+
 }
